feat: add notebook to convert to any file
# Motivation

It can be very interesting to see the resources scraped from a website
downloaded locally, so that the end-user can use them for various
purposes.

Notebook 4 is only able to write `HTML` to disk, while the notebook used
in this commit is able to write most Multipurpose Internet Mail
Extensions or `MIME` types to disk.

# Implementation

The file content is written to disk by first guessing the `MIME` type
from the content. The reason why we do this is that some websites use
the wrong `MIME` type, so it seems to be safer to detect the type using
`magic`.

If the detected type does not have a file extension, we try to get the
file extension using the provided `MIME` type instead as a last resort.

If this also fails we just give up and print a warning.